Author-Project/requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
easyocr==1.4.2
idna==3.3
imageio==2.19.2
joblib==1.1.0
networkx==2.8.2
numpy==1.22.4
opencv-python==4.5.5.64
opencv-python-headless==4.5.4.60
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
python-bidi==0.4.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
threadpoolctl==3.1.0
tifffile==2022.5.4
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
